Automatic upgrade test
This patch is a fork off of Christians automated upgrade test simplified
to focus on running a single hardcoded upgrade test flow, and modified
to run in CI.

Currently this runs the happy path test before and after a chain
upgrade, submitting and executing the upgrade proposal and then swapping
out the binaries before restarting the chain.

The design in this patch is focused on CI execution and is quite slow to
run locally so the mount based execution should be added as a variant of
the test script.

Co-authored-by: Justin Kilpatrick <justin@althea.net>